Extract StatusInfoModal to separate component file
